pulled out main functionality to new module `simclusters.py`
